% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbdDMAT-package.R
\docType{methods}
\name{Arithmetic}
\alias{*}
\alias{*,ddmatrix,ddmatrix-method}
\alias{*,ddmatrix,numeric-method}
\alias{*,numeric,ddmatrix-method}
\alias{*-method}
\alias{+}
\alias{+,ddmatrix,ddmatrix-method}
\alias{+,ddmatrix,numeric-method}
\alias{+,numeric,ddmatrix-method}
\alias{+-method}
\alias{-}
\alias{-,ddmatrix,ddmatrix-method}
\alias{-,ddmatrix,missing-method}
\alias{-,ddmatrix,numeric-method}
\alias{-,numeric,ddmatrix-method}
\alias{--method}
\alias{/}
\alias{/,ddmatrix,ddmatrix-method}
\alias{/,ddmatrix,numeric-method}
\alias{/,numeric,ddmatrix-method}
\alias{/-method}
\alias{Arithmetic}
\alias{\%/\%}
\alias{\%/\%,ddmatrix,ddmatrix-method}
\alias{\%/\%,ddmatrix,numeric-method}
\alias{\%/\%,numeric,ddmatrix-method}
\alias{\%/\%-method}
\alias{\%\%}
\alias{\%\%,ddmatrix,ddmatrix-method}
\alias{\%\%,ddmatrix,numeric-method}
\alias{\%\%,numeric,ddmatrix-method}
\alias{\%\%-method}
\alias{^}
\alias{^,ddmatrix,ddmatrix-method}
\alias{^,ddmatrix,numeric-method}
\alias{^,numeric,ddmatrix-method}
\alias{^-method}
\title{Arithmetic Operators}
\arguments{
\item{x,y}{numeric distributed matrices or numeric vectors}
}
\value{
Returns a distributed matrix.
}
\description{
Binary operations for distributed matrix/distributed matrix and distributed
matrix/vector operations.
}
\details{
If \code{x} and \code{y} are distributed matrices, then they must be
conformable, on the same BLACS context, and have the same blocking
dimension.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\", y =
\"ddmatrix\")")}{} \item{list("signature(x = \"numeric\", y =
\"ddmatrix\")")}{} \item{list("signature(x = \"ddmatrix\", y =
\"numeric\")")}{} }
}
\examples{
\dontrun{
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

# don't do this in production code
x <- matrix(1:9, 3)
x <- as.ddmatrix(x)

y <- (2*x) - x^(.5)
print(y)

finalize()
}
}
\seealso{
\code{\link{Arithmetic}, \link{LinAlg}, \link{MatMult}}
}
\keyword{Methods}

