% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_logic.r
\name{Comparators}
\alias{Comparators}
\alias{==,ddmatrix,ddmatrix-method}
\alias{!=,ddmatrix,ddmatrix-method}
\alias{all,ddmatrix-method}
\alias{any,ddmatrix-method}
\alias{<,ddmatrix,ddmatrix-method}
\alias{>,ddmatrix,ddmatrix-method}
\alias{<=,ddmatrix,ddmatrix-method}
\alias{>=,ddmatrix,ddmatrix-method}
\alias{|,ddmatrix,ddmatrix-method}
\alias{&,ddmatrix,ddmatrix-method}
\alias{<,ddmatrix,numeric-method}
\alias{<,numeric,ddmatrix-method}
\alias{>,ddmatrix,numeric-method}
\alias{>,numeric,ddmatrix-method}
\alias{<=,ddmatrix,numeric-method}
\alias{<=,numeric,ddmatrix-method}
\alias{>=,ddmatrix,numeric-method}
\alias{>=,numeric,ddmatrix-method}
\alias{==,ddmatrix,numeric-method}
\alias{==,numeric,ddmatrix-method}
\alias{!=,ddmatrix,numeric-method}
\alias{!=,numeric,ddmatrix-method}
\alias{|,ddmatrix,numeric-method}
\alias{|,numeric,ddmatrix-method}
\alias{&,ddmatrix,numeric-method}
\alias{&,numeric,ddmatrix-method}
\title{Logical Comparisons}
\usage{
\S4method{==}{ddmatrix,ddmatrix}(e1, e2)

\S4method{!=}{ddmatrix,ddmatrix}(e1, e2)

\S4method{all}{ddmatrix}(x, na.rm = FALSE)

\S4method{any}{ddmatrix}(x, na.rm = FALSE)

\S4method{<}{ddmatrix,ddmatrix}(e1, e2)

\S4method{>}{ddmatrix,ddmatrix}(e1, e2)

\S4method{<=}{ddmatrix,ddmatrix}(e1, e2)

\S4method{>=}{ddmatrix,ddmatrix}(e1, e2)

\S4method{|}{ddmatrix,ddmatrix}(e1, e2)

\S4method{&}{ddmatrix,ddmatrix}(e1, e2)

\S4method{<}{ddmatrix,numeric}(e1, e2)

\S4method{<}{numeric,ddmatrix}(e1, e2)

\S4method{>}{ddmatrix,numeric}(e1, e2)

\S4method{>}{numeric,ddmatrix}(e1, e2)

\S4method{<=}{ddmatrix,numeric}(e1, e2)

\S4method{<=}{numeric,ddmatrix}(e1, e2)

\S4method{>=}{ddmatrix,numeric}(e1, e2)

\S4method{>=}{numeric,ddmatrix}(e1, e2)

\S4method{==}{ddmatrix,numeric}(e1, e2)

\S4method{==}{numeric,ddmatrix}(e1, e2)

\S4method{!=}{ddmatrix,numeric}(e1, e2)

\S4method{!=}{numeric,ddmatrix}(e1, e2)

\S4method{|}{ddmatrix,numeric}(e1, e2)

\S4method{|}{numeric,ddmatrix}(e1, e2)

\S4method{&}{ddmatrix,numeric}(e1, e2)

\S4method{&}{numeric,ddmatrix}(e1, e2)
}
\arguments{
\item{e1, e2, x}{distributed matrix or numeric vector}

\item{na.rm}{logical, indicating whether or not \code{NA}'s should first be
removed. If not and an NA is present, \code{NA} is returned.}
}
\value{
Returns a distributed matrix.
}
\description{
Logical comparisons.
}
\details{
Performs the indicated logical comparison.

If \code{na.rm} is \code{TRUE} and only \code{NA}'s are present, then
\code{TRUE} is returned.
}
\keyword{Extraction}
\keyword{Methods}
\keyword{Type}
