% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_globalVariables.r
\docType{data}
\name{pbdDMAT Control}
\alias{dmat_opts}
\alias{pbdDMAT Control}
\title{Some default parameters for pbdDMAT.}
\format{Objects contain several parameters for communicators and methods.}
\usage{
dmat_opts
}
\description{
This set of controls is used to provide default values in this package.
}
\details{
The default blocking \code{dmat_opts$BLDIM} is \code{c(16,16)}, which results in a 16 by
16 blocking dimension for distributed matrices.  Any time a function takes
the \code{bldim=} argument, it will default to this value unless the user
specifies an alternative.

The default ICTXT is 0.  This is the full 2-dimensional processor grid.
}
\keyword{datasets}

