% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_chol.r
\name{ddmatrix-chol}
\alias{ddmatrix-chol}
\alias{chol}
\alias{chol,ddmatrix-method}
\title{Cholesky Factorization}
\usage{
\S4method{chol}{ddmatrix}(x)
}
\arguments{
\item{x}{numeric distributed matrices.}

\item{...}{Ignored.}
}
\value{
\code{chol()} performs Cholesky factorization.
}
\description{
Cholesky factorization for distributed matrices with R-like syntax, with
calculations performed by the PBLAS and ScaLAPACK libraries.
}
\details{
Extensions of R linear algebra functions.
}
\keyword{Algebra}
\keyword{Linear}
\keyword{Methods}
