% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes.r
\docType{class}
\name{ddmatrix-class}
\alias{ddmatrix-class}
\title{Class ddmatrix}
\description{
Distributed matrix class.
}
\section{Slots}{

\describe{
\item{\code{DATA}}{The local submatrix.}

\item{\code{bldim}}{Blocking factor.}

\item{\code{ICTXT}}{BLACS ICTXT value.  Should be one of 0, 1, or 2 (initialized from
\code{pbdBASE::init.grid()}) or a custom value greater than 2 (created from
\code{pbdBASE::blacs_gridinit()}).}
}}

\keyword{Classes}
