% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_norm.r
\name{ddmatrix-norm}
\alias{ddmatrix-norm}
\alias{norm}
\alias{norm,ddmatrix,ANY-method}
\title{Norm}
\usage{
\S4method{norm}{ddmatrix,ANY}(x, type = c("O", "I", "F", "M", "2"))
}
\arguments{
\item{x}{numeric distributed matrices.}

\item{type}{character. Determines which matrix norm is to be used.}
}
\value{
Returns a number.
}
\description{
Computes the norm of a distributed matrix.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\")")}{} }
}

\keyword{Algebra}
\keyword{Linear}
\keyword{Methods}
\keyword{Norm}
