% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_solve.r
\docType{methods}
\name{ddmatrix-solve}
\alias{ddmatrix-solve}
\alias{solve,ddmatrix,ANY-method}
\alias{solve,ddmatrix,ddmatrix-method}
\title{Solve}
\usage{
\S4method{solve}{ddmatrix,ANY}(a)

\S4method{solve}{ddmatrix,ddmatrix}(a, b)
}
\arguments{
\item{a, b}{numeric distributed matrices.  Here, \code{a}
and \code{b} must be on the same BLACS context and have the same blocking
dimension.}
}
\value{
\code{solve()} solves systems and performs matrix inversion when argument
\code{b=} is missing.
}
\description{
Solving linear systems and matrix inversion for distributed matrices with R-like syntax, with
calculations performed by the PBLAS and ScaLAPACK libraries.
}
\details{
Extensions of R linear algebra functions.
}
\keyword{Algebra}
\keyword{Linear}
\keyword{Methods}
