% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_det.R
\docType{methods}
\name{det}
\alias{det}
\alias{determinant.ddmatrix}
\alias{det,ddmatrix-method}
\title{det}
\usage{
\method{determinant}{ddmatrix}(x, logarithm = TRUE, ...)

\S4method{det}{ddmatrix}(x, ...)
}
\arguments{
\item{x}{numeric distributed matrices.}

\item{logarithm}{Should the logarithm of the modulus be returned?}

\item{...}{Ignored.}
}
\value{
The determinant (via \code{det()}) or an object of class \code{det} (via
\code{determinant()}).
}
\description{
Calculate the determinant via LU decomposition.
}
\details{
Regardless of the value of \code{logarithm}, the determinant is calculated
by adding the log of the diagonal of U (separately tracking the sign via
the pivot matrix).
}
\keyword{Methods}
