% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pbdDMAT-package.R
\docType{methods}
\name{Expm}
\alias{Expm}
\alias{expm}
\alias{expm,ddmatrix-method}
\alias{expm,matrix-method}
\alias{expm-method}
\title{Matrix Exponentiation}
\arguments{
\item{x}{A numeric matrix or a numeric distributed matrix.}
}
\value{
Returns a distributed matrix.
}
\description{
Routines for matrix exponentiation.
}
\details{
Formally, the exponential of a square matrix \code{X} is a power series:

\eqn{expm(X) = id + X/1! + X^2/2! + X^3/3! + \dots}

where the powers on the matrix correspond to matrix-matrix multiplications.

\code{expm()} directly computes the matrix exponential of a distributed,
dense matrix.  The implementation uses Pade' approximations and a
scaling-and-squaring technique (see references).
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\")")}{}
\item{list("signature(x = \"matrix\")")}{} }
}
\references{
Matrix exponentiation using Pade' approximations and scaling and
squaring from: "New Scaling and Squaring Algorithm for the Matrix
Exponential" Awad H. Al-Mohy and Nicholas J. Higham, August 2009
}
\seealso{
\code{\link{Arithmetic}, \link{Reductions}, \link{MatMult},
\link{LinAlg}}
}
\keyword{Algebra}
\keyword{Linear}
\keyword{Methods}

