% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pbdDMAT-package.R
\name{internal}
\alias{.conflicts.OK}
\alias{La.svd,ANY-method}
\alias{colMax,matrix-method}
\alias{colMin,matrix-method}
\alias{diag,matrix-method}
\alias{dmat.allcolreduce}
\alias{dmat.allrowreduce}
\alias{dmat.as.block}
\alias{dmat.as.blockcyclic}
\alias{dmat.as.colblock}
\alias{dmat.as.colcyclic}
\alias{dmat.as.ddmatrix}
\alias{dmat.as.rowblock}
\alias{dmat.as.rowcyclic}
\alias{dmat.blacsreduction}
\alias{dmat.bldim}
\alias{dmat.clmn}
\alias{dmat.clscl}
\alias{dmat.colreduce}
\alias{dmat.crossprod}
\alias{dmat.ddmatmult}
\alias{dmat.gmat}
\alias{dmat.ictxt}
\alias{dmat.ldim}
\alias{dmat.print}
\alias{dmat.qr.R}
\alias{dmat.rank_k}
\alias{dmat.rcminmax}
\alias{dmat.rcsum}
\alias{dmat.reblock}
\alias{dmat.redistribute}
\alias{dmat.rowreduce}
\alias{dmat.scale.center.atomic}
\alias{dmat.scale.center.ddmatrix}
\alias{dmat.scale.center.logical}
\alias{dmat.scale.scale.atomic}
\alias{dmat.scale.scale.ddmatrix}
\alias{dmat.scale.scale.logical}
\alias{dmat.submatrix}
\alias{dmat.svd}
\alias{dmat.sweep}
\alias{internal}
\alias{rowMax,matrix-method}
\alias{rowMin,matrix-method}
\alias{show,ddmatrix-method}
\alias{svd,ANY-method}
\title{All DMAT Internal Functions}
\description{
All DMAT internal functions
}
\keyword{internal}

