% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isSymmetric.r
\docType{methods}
\name{isSymmetric}
\alias{isSymmetric}
\alias{isSymmetric,ddmatrix-method}
\title{isSymmetric}
\usage{
\S4method{isSymmetric}{ddmatrix}(object, tol = 100 * .Machine$double.eps, ...)
}
\arguments{
\item{object}{Distributed matrix}

\item{tol}{Numerical tolerance for the comparison.}

\item{...}{Additional arguments passed to \code{all.equal()}.}
}
\description{
Tests if a distributed matrix is symmetric.
}
\details{
The test is performed by comparing the object against its transpose.
}

