% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pbdDMAT-package.R
\docType{methods}
\name{Summary}
\alias{Summary}
\alias{summary}
\alias{summary,ddmatrix-method}
\alias{summary-method}
\title{Distributed Matrix Summary}
\arguments{
\item{object}{numeric distributed matrix}
}
\value{
A table on processor 0, \code{NULL} on all other processors.
}
\description{
Summarize a distributed matrix.  Gives min, max, mean, etc. by column.
}
\details{
The return is on process 0 only.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\")")}{} }
}
\examples{
\dontrun{
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

# don't do this in production code
x <- matrix(1:16, ncol=4)
dx <- as.ddmatrix(x)

summary(dx)

finalize()
}
}
\keyword{Methods}

