% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_sweep.R
\name{sweep}
\alias{sweep}
\alias{sweep,ddmatrix,ANY,vector-method}
\alias{sweep,ddmatrix,ANY,ddmatrix-method}
\title{Sweep}
\usage{
sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

\S4method{sweep}{ddmatrix,ANY,vector}(x, MARGIN, STATS, FUN = "-")

\S4method{sweep}{ddmatrix,ANY,ddmatrix}(x, MARGIN, STATS, FUN = "-")
}
\arguments{
\item{x}{numeric distributed matrix.}

\item{MARGIN}{subscript over which the function will be applied}

\item{STATS}{array to be swept out.}

\item{FUN}{function used in the sweep. Only \code{+}, \code{-}, \code{*},
and \code{/} are accepted.  For more general operations, use \code{apply()}.}

\item{check.margin, ...}{Ignored.}
}
\value{
Returns a distributed matrix.
}
\description{
Sweep vector or ddmatrix from a distributed matrix.
}
\keyword{Methods}
