% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbdDMAT-package.R
\docType{methods}
\name{Type}
\alias{Type}
\alias{is.ddmatrix}
\alias{is.infinite}
\alias{is.infinite,ddmatrix-method}
\alias{is.infinite-method}
\alias{is.na}
\alias{is.na,ddmatrix-method}
\alias{is.na-method}
\alias{is.nan}
\alias{is.nan,ddmatrix-method}
\alias{is.nan-method}
\alias{is.numeric}
\alias{is.numeric,ddmatrix-method}
\alias{is.numeric-method}
\title{Type Checks, Including NA, NaN, etc.}
\arguments{
\item{x}{numeric distributed matrix}
}
\value{
Returns boolean in the case of \code{is.numeric()} and
\code{is.ddmatrix()}, otherwise a distributed matrix.
}
\description{
Functions to check for various types.
}
\details{
Performs the appropriate type check.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\")")}{} }
}
\seealso{
\code{\link{NAs}}
}
\keyword{Methods}
\keyword{Type}

