% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbdDMAT-package.R
\docType{methods}
\name{Comparators}
\alias{!=}
\alias{!=,ddmatrix,ddmatrix-method}
\alias{!=,ddmatrix,numeric-method}
\alias{!=,numeric,ddmatrix-method}
\alias{&}
\alias{&,ddmatrix,ddmatrix-method}
\alias{&,ddmatrix,numeric-method}
\alias{&,numeric,ddmatrix-method}
\alias{<}
\alias{<,ddmatrix,ddmatrix-method}
\alias{<,ddmatrix,numeric-method}
\alias{<,numeric,ddmatrix-method}
\alias{<=}
\alias{<=,ddmatrix,ddmatrix-method}
\alias{<=,ddmatrix,numeric-method}
\alias{<=,numeric,ddmatrix-method}
\alias{==}
\alias{==,ddmatrix,ddmatrix-method}
\alias{==,ddmatrix,numeric-method}
\alias{==,numeric,ddmatrix-method}
\alias{>}
\alias{>,ddmatrix,ddmatrix-method}
\alias{>,ddmatrix,numeric-method}
\alias{>,numeric,ddmatrix-method}
\alias{>=}
\alias{>=,ddmatrix,ddmatrix-method}
\alias{>=,ddmatrix,numeric-method}
\alias{>=,numeric,ddmatrix-method}
\alias{Comparators}
\alias{all}
\alias{all,ddmatrix-method}
\alias{all-method}
\alias{any}
\alias{any,ddmatrix-method}
\alias{any-method}
\alias{|}
\alias{|,ddmatrix,ddmatrix-method}
\alias{|,ddmatrix,numeric-method}
\alias{|,numeric,ddmatrix-method}
\title{Logical Comparisons}
\arguments{
\item{x,y}{distributed matrix or numeric vector}

\item{na.rm}{logical, indicating whether or not \code{NA}'s should first be
removed. If not and an NA is present, \code{NA} is returned.}
}
\value{
Returns a distributed matrix.
}
\description{
Logical comparisons.
}
\details{
Performs the indicated logical comparison.

If \code{na.rm} is \code{TRUE} and only \code{NA}'s are present, then
\code{TRUE} is returned.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\")")}{} }
}
\examples{
\dontrun{
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

# don't do this in production code
x <- matrix(sample(0, 1, 9, replace=T), 3)
comm.print(x)

x <- as.ddmatrix(x, bldim=2)

y <- any(x)
comm.print(y)

finalize()
}
}
\seealso{
\code{\link{Type}}
}
\keyword{Extraction}
\keyword{Methods}
\keyword{Type}

