% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddmatrix_norm.r
\docType{methods}
\name{ddmatrix-norm}
\alias{ddmatrix-norm}
\alias{norm}
\alias{norm,ddmatrix,ANY-method}
\title{Norm}
\usage{
\S4method{norm}{ddmatrix,ANY}(x, type = c("O", "I", "F", "M", "2"))
}
\arguments{
\item{x}{numeric distributed matrices.}

\item{type}{character. Determines which matrix norm is to be used.}
}
\value{
Returns a number.
}
\description{
Computes the norm of a distributed matrix.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\")")}{} }
}
\examples{
\dontrun{
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

comm.set.seed(diff=T)
x <- ddmatrix("rnorm", 10, 10)

nrm <- norm(x)

comm.print(nrm)

finalize()
}
}
\keyword{Algebra}
\keyword{Linear}
\keyword{Methods}
\keyword{Norm}

