% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddmatrix_solve.r
\docType{methods}
\name{ddmatrix-solve}
\alias{ddmatrix-solve}
\alias{solve,ddmatrix,ANY-method}
\alias{solve,ddmatrix,ddmatrix-method}
\title{Solve}
\usage{
\S4method{solve}{ddmatrix,ANY}(a)

\S4method{solve}{ddmatrix,ddmatrix}(a, b)
}
\arguments{
\item{a,b}{numeric distributed matrices.  Here, \code{a}
and \code{b} must be on the same BLACS context and have the same blocking
dimension.}
}
\value{
\code{solve()} solves systems and performs matrix inversion when argument
\code{b=} is missing.
}
\description{
Solving linear systems and matrix inversion for distributed matrices with R-like syntax, with
calculations performed by the PBLAS and ScaLAPACK libraries.
}
\details{
Extensions of R linear algebra functions.
}
\examples{
\dontrun{
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

x <- ddmatrix(1:9, 3)

y <- solve(t(A) \%*\% A)
print(y)

finalize()
}
}
\keyword{Algebra}
\keyword{Linear}
\keyword{Methods}

