% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_summary.r
\docType{methods}
\name{ddmatrix-summary}
\alias{ddmatrix-summary}
\alias{summary}
\alias{summary,ddmatrix-method}
\title{Distributed Matrix Summary}
\usage{
summary(object, ...)

\S4method{summary}{ddmatrix}(object)
}
\arguments{
\item{object}{numeric distributed matrix}

\item{...}{Additional arguments.}
}
\value{
A table on processor 0, \code{NULL} on all other processors.
}
\description{
Summarize a distributed matrix.  Gives min, max, mean, etc. by column.
}
\details{
The return is on process 0 only.
}
\keyword{Methods}
