% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_isdot.r
\name{isdot}
\alias{isdot}
\alias{is.ddmatrix}
\alias{is.na,ddmatrix-method}
\alias{is.nan,ddmatrix-method}
\alias{is.numeric,ddmatrix-method}
\alias{is.infinite,ddmatrix-method}
\title{Type Checks, Including NA, NaN, etc.}
\usage{
is.ddmatrix(x)

\S4method{is.na}{ddmatrix}(x)

\S4method{is.nan}{ddmatrix}(x)

\S4method{is.numeric}{ddmatrix}(x)

\S4method{is.infinite}{ddmatrix}(x)
}
\arguments{
\item{x}{numeric distributed matrix}
}
\value{
Returns boolean in the case of \code{is.numeric()} and
\code{is.ddmatrix()}, otherwise a distributed matrix.
}
\description{
Functions to check for various types.
}
\details{
Performs the appropriate type check.
}
\keyword{Methods}
\keyword{Type}
