% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pbdDMAT-package.R
\docType{methods}
\name{kappa}
\alias{ConditionNumbers}
\alias{kappa}
\alias{kappa.ddmatrix}
\alias{rcond}
\alias{rcond,ddmatrix-method}
\alias{rcond-method}
\title{Compute or estimate the Condition Number of a Distributed Matrix}
\arguments{
\item{x,z}{numeric distributed matrices.}

\item{exact}{logical. Determines whether exact condition number or
approximation should be computed.}

\item{norm}{character. Determines which matrix norm is to be used.}

\item{method}{character. Determines the method use in computing condition
number.}

\item{triangular}{logical. If true, only the lower triangle is used.}

\item{...}{Extra arguments.}
}
\value{
Returns a number.
}
\description{
Computes or estimates the condition number.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\")")}{}
\item{list("signature(z = \"ddmatrix\")")}{} }
}
\examples{
\dontrun{
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

comm.set.seed(diff=T)
x <- ddmatrix("rnorm", 10, 10)

cnm <- rcond(x)

comm.print(cnm)

finalize()
}
}
\seealso{
\code{\link{Norm}}
}
\keyword{Algebra}
\keyword{ConditionNumbers}
\keyword{Linear}
\keyword{Methods}

