% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddmatrix_math.R
\docType{methods}
\name{MiscMath}
\alias{MiscMath}
\alias{abs}
\alias{abs,ddmatrix-method}
\alias{abs-method}
\alias{acos}
\alias{acos,ddmatrix-method}
\alias{acos-method}
\alias{asin}
\alias{asin,ddmatrix-method}
\alias{asin-method}
\alias{atan}
\alias{atan,ddmatrix-method}
\alias{atan-method}
\alias{cos}
\alias{cos,ddmatrix-method}
\alias{cos-method}
\alias{cosh}
\alias{cosh,ddmatrix-method}
\alias{cosh-method}
\alias{exp}
\alias{exp,ddmatrix-method}
\alias{exp-method}
\alias{log}
\alias{log,ddmatrix-method}
\alias{log-method}
\alias{log10}
\alias{log10,ddmatrix-method}
\alias{log10-method}
\alias{log1p}
\alias{log1p,ddmatrix-method}
\alias{log1p-method}
\alias{log2}
\alias{log2,ddmatrix-method}
\alias{log2-method}
\alias{sin}
\alias{sin,ddmatrix-method}
\alias{sin-method}
\alias{sinh}
\alias{sinh,ddmatrix-method}
\alias{sinh-method}
\alias{sqrt}
\alias{sqrt,ddmatrix-method}
\alias{sqrt-method}
\alias{tan}
\alias{tan,ddmatrix-method}
\alias{tan-method}
\alias{tanh}
\alias{tanh,ddmatrix-method}
\alias{tanh-method}
\title{Miscellaneous Mathematical Functions}
\usage{
\S4method{sqrt}{ddmatrix}(x)

\S4method{abs}{ddmatrix}(x)

\S4method{exp}{ddmatrix}(x)

\S4method{log}{ddmatrix}(x, base = exp(1))

\S4method{log2}{ddmatrix}(x)

\S4method{log10}{ddmatrix}(x)

\S4method{log1p}{ddmatrix}(x)

\S4method{sin}{ddmatrix}(x)

\S4method{cos}{ddmatrix}(x)

\S4method{tan}{ddmatrix}(x)

\S4method{asin}{ddmatrix}(x)

\S4method{acos}{ddmatrix}(x)

\S4method{atan}{ddmatrix}(x)

\S4method{sinh}{ddmatrix}(x)

\S4method{cosh}{ddmatrix}(x)

\S4method{tanh}{ddmatrix}(x)
}
\arguments{
\item{x}{numeric distributed matrix}

\item{base}{a positive number: the base with respect to which logarithms are
computed. Defaults to e='exp(1)'.}
}
\value{
Returns a distributed matrix.
}
\description{
Binary operations for distributed matrix/distributed matrix and distributed
matrix/vector operations.
}
\details{
Performs the miscellaneous mathematical calculation on a distributed matrix.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\")")}{} }
}
\examples{
\dontrun{
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

# don't do this in production code
x <- matrix(1:9, 3)
x <- as.ddmatrix(x)

y <- sqrt(abs(log(x/10)))
comm.print(y)

finalize()
}
}
\seealso{
\code{\link{Arithmetic}, \link{Reductions}}
}
\keyword{Methods}

