% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_math.R
\docType{methods}
\name{math}
\alias{math}
\alias{sqrt,ddmatrix-method}
\alias{abs,ddmatrix-method}
\alias{exp,ddmatrix-method}
\alias{log,ddmatrix-method}
\alias{log2,ddmatrix-method}
\alias{log10,ddmatrix-method}
\alias{log1p,ddmatrix-method}
\alias{sin,ddmatrix-method}
\alias{cos,ddmatrix-method}
\alias{tan,ddmatrix-method}
\alias{asin,ddmatrix-method}
\alias{acos,ddmatrix-method}
\alias{atan,ddmatrix-method}
\alias{sinh,ddmatrix-method}
\alias{cosh,ddmatrix-method}
\alias{tanh,ddmatrix-method}
\title{Miscellaneous Mathematical Functions}
\usage{
\S4method{sqrt}{ddmatrix}(x)

\S4method{abs}{ddmatrix}(x)

\S4method{exp}{ddmatrix}(x)

\S4method{log}{ddmatrix}(x, base = exp(1))

\S4method{log2}{ddmatrix}(x)

\S4method{log10}{ddmatrix}(x)

\S4method{log1p}{ddmatrix}(x)

\S4method{sin}{ddmatrix}(x)

\S4method{cos}{ddmatrix}(x)

\S4method{tan}{ddmatrix}(x)

\S4method{asin}{ddmatrix}(x)

\S4method{acos}{ddmatrix}(x)

\S4method{atan}{ddmatrix}(x)

\S4method{sinh}{ddmatrix}(x)

\S4method{cosh}{ddmatrix}(x)

\S4method{tanh}{ddmatrix}(x)
}
\arguments{
\item{x}{numeric distributed matrix}

\item{base}{a positive number: the base with respect to which logarithms are
computed. Defaults to e='exp(1)'.}
}
\value{
Returns a distributed matrix.
}
\description{
Binary operations for distributed matrix/distributed matrix and distributed
matrix/vector operations.
}
\details{
Performs the miscellaneous mathematical calculation on a distributed matrix.
}
\keyword{Methods}
