% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbdDMAT-package.R
\docType{methods}
\name{Reductions}
\alias{Reductions}
\alias{colMax}
\alias{colMax,ddmatrix-method}
\alias{colMax-method}
\alias{colMeans}
\alias{colMeans,ddmatrix-method}
\alias{colMeans-method}
\alias{colMin}
\alias{colMin,ddmatrix-method}
\alias{colMin-method}
\alias{colSums}
\alias{colSums,ddmatrix-method}
\alias{colSums-method}
\alias{max}
\alias{max,ddmatrix-method}
\alias{max-method}
\alias{mean}
\alias{mean,ddmatrix-method}
\alias{mean-method}
\alias{median}
\alias{median,ddmatrix-method}
\alias{median-method}
\alias{min}
\alias{min,ddmatrix-method}
\alias{min-method}
\alias{prod}
\alias{prod,ddmatrix-method}
\alias{prod-method}
\alias{rowMax}
\alias{rowMax,ddmatrix-method}
\alias{rowMax-method}
\alias{rowMeans}
\alias{rowMeans,ddmatrix-method}
\alias{rowMeans-method}
\alias{rowMin}
\alias{rowMin,ddmatrix-method}
\alias{rowMin-method}
\alias{rowSums}
\alias{rowSums,ddmatrix-method}
\alias{rowSums-method}
\alias{sum}
\alias{sum,ddmatrix-method}
\alias{sum-method}
\title{Arithmetic Reductions: Sums, Means, and Prods}
\arguments{
\item{x}{numeric distributed matrix}

\item{na.rm}{logical. Should missing (including \code{NaN}) be removed?}

\item{...}{additional arguments}
}
\value{
Returns a global numeric vector.
}
\description{
Arithmetic reductions for distributed matrices.
}
\details{
Performs the reduction operation on a distributed matrix.

There are four legitimately new operations, namely \code{rowMin()},
\code{rowMax()}, \code{colMin()}, and \code{colMax()}.  These
implementations are not really necessary in R because one can easily (and
reasonably efficiently) do something like

\code{apply(X=x, MARGIN=1L, FUN=min, na.rm=TRUE)}

But \code{apply()} on a \code{ddmatrix} is \emph{very} costly, and should be
used sparingly.
}
\section{Methods}{
 \describe{ \item{list("signature(x = \"ddmatrix\")")}{} }
}
\examples{
\dontrun{
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

# don't do this in production code
x <- matrix(1:9, 3)
x <- as.ddmatrix(x)

y <- sum(colMeans(x))
comm.print(y)

finalize()
}
}
\seealso{
\code{\link{Arithmetic}}
}
\keyword{Methods}

