% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_round.r
\docType{methods}
\name{rounding}
\alias{rounding}
\alias{round,ddmatrix-method}
\alias{ceiling,ddmatrix-method}
\alias{floor,ddmatrix-method}
\title{Rounding of Numbers}
\usage{
\S4method{round}{ddmatrix}(x, digits = 0)

\S4method{ceiling}{ddmatrix}(x)

\S4method{floor}{ddmatrix}(x)
}
\arguments{
\item{x}{numeric distributed matrix}

\item{digits}{integer indicating the number of decimal places
(\code{round()}) or significant digits (\code{signif()}) to be used.
Negative values are allowed (see 'Details').}
}
\value{
Returns a distributed matrix.
}
\description{
Extensions of R rounding functions for distributed matrices.
}
\details{
Rounding to a negative number of digits means rounding to a power of ten, so
for example \code{round(x, digits = -2)} rounds to the nearest hundred.
}
\keyword{Methods}
