% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmatrix_transpose.r
\name{transpose}
\alias{transpose}
\alias{t}
\alias{t,ddmatrix-method}
\title{Distributed Matrix Transpose}
\usage{
t(x)

\S4method{t}{ddmatrix}(x)
}
\arguments{
\item{x}{numeric distributed matrix.}
}
\value{
The transposed matrix.
}
\description{
Transposes a distributed dense matrix.
}
\keyword{Algebra}
\keyword{Linear}
\keyword{Methods}
