\name{dmat.internal}
\alias{dmat.rank_k}
\alias{dmat.sweep}


\alias{dmat.allcolreduce}
\alias{dmat.allrowreduce}

\alias{dmat.as.ddmatrix}
\alias{dmat.blacsreduction}
\alias{dmat.clmn}
\alias{dmat.clscl}

\alias{dmat.colreduce}
\alias{dmat.crossprod}
\alias{dmat.ddmatmult}
\alias{dmat.print}

\alias{dmat.qr.R}
\alias{dmat.rcsum}
\alias{dmat.rowreduce}
\alias{dmat.scale.center.atomic}

\alias{dmat.scale.center.ddmatrix}
\alias{dmat.scale.center.logical}

\alias{dmat.scale.scale.atomic}
\alias{dmat.scale.scale.ddmatrix}

\alias{dmat.scale.scale.logical}
\alias{dmat.svd}



\alias{svd,ANY-method}
\alias{La.svd,ANY-method}

\alias{.conflicts.OK}

\title{ All DMAT Internal Functions }
\description{ All DMAT internal functions }
\keyword{internal}

